Split out useCustomThemes hook (DH-19307)
